Moved Notification functions to the viewModel and fixed/updated all conflicts
